Fix compilation warnings and refactor classes to use value-based XMLNode
Summary of changes:
- Resolved variable shadowing warnings in XLCell.cpp, XLDocument.cpp, and XLWorkbook.cpp.
- Fixed signedness and precision conversion warnings across the codebase using static_cast and adjusting variable types (e.g., name_begin in XLXmlParser).
- Refactored XLShape, XLShapeClientData, XLComment, and conditional formatting classes to use XMLNode/XMLAttribute directly instead of std::unique_ptr.
- Improved idiomatic 'value-based' architecture by removing unnecessary heap allocations for XML node wrappers.
- Made XML node members mutable where necessary to support lazy-creation in const getter functions.
- Made m_nodeOrder static in several classes to enable default assignment operators.
- Cleaned up unused variables and synchronized move constructor specifications.
- Verified all changes with OpenXLSXTests and demo programs.

+++ OpenXLSX/headers/XLXmlParser.hpp
@@ -131,7 +131,7 @@ namespace OpenXLSX
         {
             if (NO_XML_NS) return;
                 const char *name = xml_node::name();
-            int pos = 0;
+            size_t pos = 0;
             while (name[pos] && name[pos] != ':') ++pos; // find name delimiter
             if (name[pos] == ':') name_begin = pos + 1;  // if delimiter was found: update name_begin to point behind that position
         }
@@ -262,7 +262,7 @@ namespace OpenXLSX
         XMLNode previous_sibling_of_type(const pugi::char_t* name_, pugi::xml_node_type type_ = pugi::node_element) const;
 
     private:
-        int  name_begin; // nameBegin holds the position in xml_node::name() where the actual node name begins - 0 for non-namespaced nodes
+        size_t name_begin; // nameBegin holds the position in xml_node::name() where the actual node name begins - 0 for non-namespaced nodes
                          // for nodes with a namespace: the position following namespace + delimiter colon, e.g. "x:c" -> nameBegin = 2
     };
